Aba quersumme

+++ Andreas Bachmann/Quersumme-Querprodukt/src/ch/stoecklin/Digits.java
@@ -4,11 +4,11 @@ public class Digits {
 
 	
 	public static Result calc(String eingabeString) {
-		int result1 = 0;
-		int result2 = 1;
+		long result1 = 0;
+		long result2 = 1;
 		for(int i = 0; i < eingabeString.length(); i++) {
-			result1 = result1 + Integer.parseInt(eingabeString.substring(i,i+1));
-			result2 = result2 * Integer.parseInt(eingabeString.substring(i,i+1));
+			result1 = result1 + Long.parseLong(eingabeString.substring(i,i+1));
+			result2 = result2 * Long.parseLong(eingabeString.substring(i,i+1));
 		}
 		Result result = new Result(result1, result2);
 		return result;
@@ -16,9 +16,9 @@ public class Digits {
 }
 
 class Result{
-	int result1;
-	int result2;
-	public Result(int result1, int result2) {
+	long result1;
+	long result2;
+	public Result(long result1, long result2) {
 		this.result1 = result1;
 		this.result2 = result2;
 	}
